+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Changelog
 
 * [Network] Allow the network interface to be specified using Network::setNetworkInterface().
 
+### Changed
+
+* [Network] Correct the cache lookup to only use cache from the same network interface and multicast address.
+
 --------
 
 ## 1.7.4 - 2015-12-03